requirements.txt
opencv-python
ultralytics
python-dotenv
roboflow